Create factory method to easily instantiate from both entry points
