+++ Wikipedia/Code/ExploreViewController.swift
@@ -66,8 +66,15 @@ class ExploreViewController: ColumnarCollectionViewController, ExploreCardViewCo
         #endif
         
         DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) { [weak self] in
-            let viewModel = WKFeatureAnnouncementViewModel(title: "Try 'Add an image'", body: "Decide if an image gets added to a Wikipedia article. You can find the ‘Add an image’ card in your ‘Explore feed’.", primaryButtonTitle: "Try now", image:  WKIcon.checkPhoto, primaryButtonAction: {
-                print("Push to image recs.")
+            let viewModel = WKFeatureAnnouncementViewModel(title: "Try 'Add an image'", body: "Decide if an image gets added to a Wikipedia article. You can find the ‘Add an image’ card in your ‘Explore feed’.", primaryButtonTitle: "Try now", image:  WKIcon.checkPhoto, primaryButtonAction: { [weak self] in
+                
+                guard let self,
+                let imageRecommendationViewController = WKImageRecommendationsViewController.imageRecommendationsViewController(dataStore: self.dataStore, imageRecDelegate: self) else {
+                    return
+                }
+                
+                navigationController?.pushViewController(imageRecommendationViewController, animated: true)
+                
             })
             self?.announceFeature(viewModel: viewModel)
         }
